committed factory simple and decorator extras

+++ Vjezbe/src/pckg_adapters_and_facades/pckg_zad1/WriterAdapter.java
@@ -10,7 +10,7 @@ public class WriterAdapter implements WriterInt {
 	
 	@Override
 	public void writeNovel ( int novelType ) {
-		if ( novelType == ( int ) novelType ) {
+		if ( novelType == novelType ) {
 			oNov.writeSFNovel ( );
 		}
 		else {
